Autoconverted sources for ME 2.8 Used sharpen ydanila/sharpen_imazen@6155275c08dc032e51f28649fbc52500f5508b46

+++ Com.Drew/Com/Drew/Metadata/Directory.cs
@@ -886,7 +886,7 @@ namespace Com.Drew.Metadata
 			}
 			if (o is int?)
 			{
-				return new Rational((long)o, 1);
+				return new Rational((int)o, 1);
 			}
 			if (o is long?)
 			{
